requirements.txt
et-xmlfile==1.1.0
loguru==0.7.2
openpyxl==3.1.3
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
PyYAML==6.0.1
